+++ src/bin/ir_cut.c
@@ -10,9 +10,7 @@ struct CommandLineArguments
 {
     unsigned int enable;
     unsigned int disable;
-    unsigned int status;
-    unsigned int json;
-} cli = {0, 0, 0, 0};
+} cli = {0, 0};
 
 
 int main(int argc, char *argv[])
@@ -21,8 +19,6 @@ int main(int argc, char *argv[])
     struct poptOption po[] = {
         {"enable",  'e', POPT_ARG_NONE, &cli.enable,  0, "Enable the IR led",         "Enable"},
         {"disable", 'd', POPT_ARG_NONE, &cli.disable, 0, "Disable the IR led",        "Disable"},
-        {"status",  's', POPT_ARG_NONE, &cli.status,  0, "Retrieve the status",       "Led Status"},
-        {"json",    'j', POPT_ARG_NONE, &cli.json,    0, "Retrieve the info in json", "Status Json"},
         POPT_AUTOHELP
         {NULL}
     };
@@ -44,12 +40,12 @@ int main(int argc, char *argv[])
         return 1;
     }
 
-    if (!cli.enable && !cli.disable && !cli.status && !cli.json) {
+    if (!cli.enable && !cli.disable) {
         poptPrintUsage(pc, stderr, 0);
         exit(1);
     }
 
-    if ((cli.enable + cli.disable + cli.status + cli.json > 1)) {
+    if ((cli.enable + cli.disable > 1)) {
         poptPrintUsage(pc, stderr, 0);
         exit(1);
     }
@@ -65,10 +61,6 @@ int main(int argc, char *argv[])
         success = ircut_on();
     else if (cli.disable)
         success = ircut_off();
-    else if (cli.status)
-        success = ircut_status();
-    else if (cli.json)
-        success = ircut_status_json();
     else {
         poptPrintUsage(pc, stderr, 0);
         exit(1);
