+++ src/lib/chuangmi_ircut.h
@@ -6,9 +6,7 @@
 
 int ircut_init(void);
 int ircut_end(void);
-int ircut_is_initialized(void);
 int ircut_status(void);
-int ircut_status_json(void);
 int ircut_on(void);
 int ircut_off(void);
