+++ src/lib/chuangmi_isp328.c
@@ -203,37 +203,6 @@ int nightmode_status(void)
 }
 
 /*
- * Print the light values
- */
-int nightmode_info(void)
-{
-    if (nightmode_update_values() < 0) {
-        fprintf(stderr, "Failed to retrieve EV and IR values\n");
-        return -1;
-    }
-
-    fprintf(stdout, "EV Value: %d\n", light_info.ev);
-    fprintf(stdout, "IR Value: %d\n", light_info.ir);
-
-    return 0;
-}
-
-/*
- * Print the light values in json
- */
-int nightmode_info_json(void)
-{
-    if (nightmode_update_values() < 0) {
-        fprintf(stderr, "Failed to retrieve EV and IR values\n");
-        return -1;
-    }
-
-    printf("{\"ev\":%d,\"ir\":%d}", light_info.ev, light_info.ir);
-
-    return 0;
-}
-
-/*
  * Set flip to a value (Valid: 1/0)
  */
 int flipmode_set(int value)
@@ -281,417 +250,6 @@ int flipmode_status(void)
     }
 
     fprintf(stdout, "%s\n", (mode == 1) ? "on" : "off");
-
-    return 0;
-}
-
-int brightness_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set brightness to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting brightness to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_BRIGHTNESS, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for brightness!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int brightness_reset(void)
-{
-    if (brightness_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset brightness to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting brightness has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int brightness_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_BRIGHTNESS, &value);
-    if (ret < 0)
-        return -1;
-
-    return value;
-}
-
-int brightness_print(void)
-{
-    int value = brightness_get();
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for brightness!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int contrast_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set contrast to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting contrast to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_CONTRAST, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for contrast!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int contrast_reset(void)
-{
-    if (contrast_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset contrast to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting contrast has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int contrast_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_CONTRAST, &value);
-
-    if (ret < 0)
-        return -1;
-
-    return value;
-}
-
-int contrast_print(void)
-{
-    int value = contrast_get();
-
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for contrast!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int hue_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set hue to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting hue to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_HUE, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for hue!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int hue_reset(void)
-{
-    if (hue_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset hue to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting hue has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int hue_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_HUE, &value);
-    if (ret < 0) {
-        return -1;
-    }
-    return value;
-}
-
-
-int hue_print(void)
-{
-    int value = hue_get();
-
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for hue!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int saturation_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set saturation to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting saturation to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_SATURATION, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for saturation!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int saturation_reset(void)
-{
-    if (saturation_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset saturation to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting saturation has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int saturation_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_SATURATION, &value);
-    if (ret < 0)
-        return -1;
-
-    return value;
-}
-
-int saturation_print(void)
-{
-    int value = saturation_get();
-
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for saturation!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int denoise_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set denoise to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting denoise to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_DENOISE, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for denoise!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int denoise_reset(void)
-{
-    if (denoise_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset denoise to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting denoise has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int denoise_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_DENOISE, &value);
-    if (ret < 0)
-        return -1;
-
-    return value;
-}
-
-int denoise_print(void)
-{
-    int value = denoise_get();
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for denoise!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int sharpness_set(int value)
-{
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    if (value < 0 && value > 255) {
-        fprintf(stderr, "Error: Cannot set sharpness to %d: Use: (0-255)\n", value);
-        return -1;
-    }
-
-    fprintf(stderr, "Setting sharpness to %d\n", value);
-
-    int ret = ioctl(isp_fd, ISP_IOC_SET_SHARPNESS, &value);
-    if (ret < 0) {
-        fprintf(stderr, "Error: Cannot set value for sharpness!\n");
-        return -1;
-    }
-
-    return 0;
-}
-
-int sharpness_reset(void)
-{
-    if (sharpness_set(128) < 0 ) {
-        fprintf(stderr, "Error: Failed to reset sharpness to it's default value!\n");
-        return -1;
-    } else {
-        fprintf(stderr, "Setting sharpness has been reset to it's default value (128)\n");
-        return 0;
-    }
-}
-
-int sharpness_get(void)
-{
-    int value;
-
-    if (isp328_is_initialized() < 0)
-        return -1;
-
-    int ret = ioctl(isp_fd, ISP_IOC_GET_SHARPNESS, &value);
-    if (ret < 0) {
-        return -1;
-    }
-
-    return value;
-}
-
-int sharpness_print(void)
-{
-    int value = sharpness_get();
-
-    if (value < 0) {
-        fprintf(stderr, "Error: Cannot get value for sharpness!\n");
-        return -1;
-    }
-
-    fprintf(stdout, "%d\n", value);
-    return 0;
-}
-
-int print_camera_info_json(void)
-{
-        fprintf(stdout, "{");
-        fprintf(stdout, "\"brightness\":%d,", brightness_get());
-        fprintf(stdout, "\"contrast\":%d,", contrast_get());
-        fprintf(stdout, "\"hue\":%d,", hue_get());
-        fprintf(stdout, "\"saturation\":%d,", saturation_get());
-        fprintf(stdout, "\"denoise\":%d,", denoise_get());
-        fprintf(stdout, "\"sharpness\":%d", sharpness_get());
-        fprintf(stdout, "}");
-
-        return 0;
-}
-
-
-int print_camera_info_shell(void)
-{
-        fprintf(stdout, "BRIGHTNESS=\"%d\"\n", brightness_get());
-        fprintf(stdout, "CONTRAST=\"%d\"\n",   contrast_get());
-        fprintf(stdout, "HUE=\"%d\"\n",        hue_get());
-        fprintf(stdout, "SATURATION=\"%d\"\n", saturation_get());
-        fprintf(stdout, "DENOISE=\"%d\"\n",    denoise_get());
-        fprintf(stdout, "SHARPNESS=\"%d\"\n",   sharpness_get());
-        return 0;
-}
-
-
-int print_camera_info(void)
-{
-        fprintf(stdout, "Settings:\n");
-        fprintf(stdout, "- Brightness: %d\n", brightness_get());
-        fprintf(stdout, "- Contrast:   %d\n", contrast_get());
-        fprintf(stdout, "- Hue:        %d\n", hue_get());
-        fprintf(stdout, "- Saturation: %d\n", saturation_get());
-        fprintf(stdout, "- Denoise:    %d\n", denoise_get());
-        fprintf(stdout, "- Sharpness:  %d\n", sharpness_get());
-        fprintf(stdout, "\n");
-
-        return 0;
-}
-
-int reset_camera_adjustments(void)
-{
-    fprintf(stderr, "Resetting all settings to default!\n");
-
-    if (brightness_reset()  < 0 ||
-        contrast_reset()    < 0 ||
-        hue_set(128)        < 0 ||
-        saturation_set(128) < 0 ||
-        denoise_set(128)    < 0 ||
-        sharpness_set(128)  < 0)
-            return -1;
 
     return 0;
 }
