+++ src/lib/chuangmi_utils.h
@@ -11,9 +11,6 @@
 #define GPIO_DIRECTION_MAX 35
 #define GPIO_VALUE_MAX 30
 
-int write_file(const char *file_path, char *content);
-int read_int(const char *filename);
-
 int gpio_export(int pin);
 int gpio_unexport(int pin);
 int gpio_direction(int pin, int dir);
